+++ __mocks__/fileMock.js
@@ -1,9 +1,7 @@
+const path = require("path");
+
 module.exports = {
-  process() {
-    return "module.exports = '';";
-  },
-  getCacheKey() {
-    // The output is always the same.
-    return "svgTransform";
+  process(src, filename) {
+    return `module.exports = ${JSON.stringify(path.basename(filename))};`;
   }
 };