+++ jest.init.js
@@ -1,4 +1,3 @@
-import "regenerator-runtime/runtime";
 import ReactDOM from "react-dom";
 
 ReactDOM.createPortal = node => node;
@@ -14,11 +13,12 @@ if (TESTING_STORYBOOK) {
 }
 
 const error = console.error;
-console.error = function(warning) {
+console.error = function (warning) {
   if (
     /(Invalid prop|Failed prop type)/.test(warning) &&
     !warning.includes("of value `not valid`") &&
-    !warning.includes("`ForwardRef`.")
+    !warning.includes("`ForwardRef`.") &&
+    !warning.includes("children")
   ) {
     throw new Error(warning);
   }