+++ src/components/ListTitle/__stories__/listTitle.stories.js
@@ -1,20 +1,9 @@
 import React from "react";
-import { action } from "@storybook/addon-actions";
-import { text, boolean, number, select } from "@storybook/addon-knobs";
 import { withPerformance } from "storybook-addon-performance";
 import ListTitle from "../ListTitle";
-import {
-  StoryStateRow,
-  StoryStateColumn,
-  ComponentStateDescription,
-  FlexLayout,
-  Divider
-} from "../../storybook-helpers";
-import DarkThemeContainer from "../../../StoryBookComponents/DarkThemeContainer/DarkThemeContainer";
-import StoryWrapper from "../../../StoryBookComponents/StoryWrapper/StoryWrapper";
 
 export const Sandbox = () => {
-  const title = text("Text", "Essentials");
+  const title = "Essentials";
   return (
     <div>
       <ListTitle id="Knobs">{title}</ListTitle>