+++ src/components/Menu/Menu/Menu.jsx
@@ -31,7 +31,8 @@ const Menu = forwardRef(
       focusItemIndex,
       focusItemIndexOnMount,
       isSubMenu,
-      useDocumentEventListeners
+      useDocumentEventListeners,
+      shouldScrollMenu
     },
     forwardedRef
   ) => {
@@ -68,13 +69,8 @@ const Menu = forwardRef(
       });
     }, [originalChildren]);
 
-    const {
-      setSubMenuIsOpenByIndex,
-      hasOpenSubMenu,
-      openSubMenuIndex,
-      setOpenSubMenuIndex,
-      resetOpenSubMenuIndex
-    } = useSubMenuIndex();
+    const { setSubMenuIsOpenByIndex, hasOpenSubMenu, openSubMenuIndex, setOpenSubMenuIndex, resetOpenSubMenuIndex } =
+      useSubMenuIndex();
 
     const onCloseMenu = useOnCloseMenu(onSetActiveItemIndexCallback, setOpenSubMenuIndex, onClose);
 
@@ -151,7 +147,8 @@ const Menu = forwardRef(
                   closeMenu: onCloseMenu,
                   menuId: id,
                   useDocumentEventListeners,
-                  isInitialSelectedState
+                  isInitialSelectedState,
+                  shouldScrollMenu
                 })
               : null;
           })}
@@ -177,7 +174,8 @@ Menu.defaultProps = {
   focusItemIndex: -1,
   isSubMenu: false,
   useDocumentEventListeners: false,
-  focusItemIndexOnMount: -1
+  focusItemIndexOnMount: -1,
+  shouldScrollMenu: false
 };
 
 Menu.propTypes = {
@@ -193,7 +191,8 @@ Menu.propTypes = {
   focusItemIndex: PropTypes.number,
   isSubMenu: PropTypes.bool,
   useDocumentEventListeners: PropTypes.bool,
-  focusItemIndexOnMount: PropTypes.number
+  focusItemIndexOnMount: PropTypes.number,
+  shouldScrollMenu: PropTypes.bool
 };
 
 export default Menu;